Fix Rotation problem (No Need Two Trail)
Trail can operate on a single axis since it is not given a rotation depending on the direction it goes. The same trail objects are added for two axes and this leads to performance losses. To solve this, I added a code that changes the rotation of the trail according to direction.